% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_calculate_interval.R
\name{bias_validation_interval}
\alias{bias_validation_interval}
\title{Calculate bias validation interval}
\usage{
bias_validation_interval(TV, m, se_c)
}
\arguments{
\item{TV}{True value}

\item{m}{factor}

\item{se_c}{SE Combined}
}
\value{
named list with the interval
}
\description{
Calculate bias validation interval
}
