% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_calculate_aov_info.R
\name{calculate_aov_infos}
\alias{calculate_aov_infos}
\title{Calculate ANOVA Results and Imprecision Estimates}
\usage{
calculate_aov_infos(ep_15_table)
}
\arguments{
\item{ep_15_table}{table generated from create_table_ep_15()}
}
\value{
Named list with ANOVA Results and Imprecision Estimates
}
\description{
Calculate ANOVA Results and Imprecision Estimates
}
\examples{
calculate_aov_infos(create_table_ep_15(CLSIEP15::ferritin_long, data_type = 'long'))
}
