% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_df_combined.R
\name{calculate_df_combined}
\alias{calculate_df_combined}
\title{Calculate degrees of freedom of SE C (SE combined) given a selected scenario and additional parameters necessary for the scenario}
\usage{
calculate_df_combined(scenario, ...)
}
\arguments{
\item{scenario}{Scenario (A, B, C, D, E)}

\item{...}{additional parameters necessary for the scenario}
}
\value{
DF
}
\description{
Calculate degrees of freedom of SE C (SE combined) given a selected scenario and additional parameters necessary for the scenario
}
