% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_calculate_uvl_info.R
\name{calculate_dfWL}
\alias{calculate_dfWL}
\title{Calculate degres of freedom within-lab as specified in appendix B}
\usage{
calculate_dfWL(cvr_manufacture, cvwl_manufacture, k, n0, N)
}
\arguments{
\item{cvr_manufacture}{CV repeatability informed by the manufacturer}

\item{cvwl_manufacture}{CV within-lab informed by the manufacturer}

\item{k}{the number of runs}

\item{n0}{the “average” number of results per run}

\item{N}{the total number of replicates}
}
\value{
dfwl
}
\description{
Calculate degres of freedom within-lab as specified in appendix B
}
