% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_calculate_uvl_info.R
\name{calculate_F_uvl}
\alias{calculate_F_uvl}
\title{Calculate the UVL factor}
\usage{
calculate_F_uvl(nsamp = 1, df, alpha = 0.05)
}
\arguments{
\item{nsamp}{n samples in the study}

\item{df}{degres of freedom}

\item{alpha}{confidence level}
}
\value{
Uvl factor
}
\description{
Calculate the UVL factor
}
