% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_m.R
\name{calculate_m}
\alias{calculate_m}
\title{Calculate M}
\usage{
calculate_m(df, conf.level = 95, nsamples = 1)
}
\arguments{
\item{df}{degrees of freedom}

\item{conf.level}{confidence interval}

\item{nsamples}{number of samples}
}
\value{
m factor
}
\description{
Calculate M
}
