% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_calculate_aov_info.R
\name{calculate_n0}
\alias{calculate_n0}
\title{Calculate n0}
\usage{
calculate_n0(long_result_table)
}
\arguments{
\item{long_result_table}{table generated by create_table_ep_15 function}
}
\value{
The n0 number which refers to Number of Results per Run
}
\description{
Calculate n0
}
