% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_se_c.R
\name{calculate_se_c}
\alias{calculate_se_c}
\title{Calculate SE combined based on SE X and SE RM}
\usage{
calculate_se_c(se_x, se_rm)
}
\arguments{
\item{se_x}{SE X}

\item{se_rm}{SE RM}
}
\value{
SE C
}
\description{
Calculate SE combined based on SE X and SE RM
}
