% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_se_rm.R
\name{calculate_se_rm}
\alias{calculate_se_rm}
\title{Calculate SE RM given a scenario and a list of additional args that can change based on the selected scenario or sub scenario}
\usage{
calculate_se_rm(scenario, additional_args)
}
\arguments{
\item{scenario}{scenario (A, B, C, D, E)}

\item{additional_args}{additional arguments list}
}
\value{
SE RM
}
\description{
Calculate SE RM given a scenario and a list of additional args that can change based on the selected scenario or sub scenario
}
