% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_se_rm.R
\name{calculate_se_rm_a_lowerupper}
\alias{calculate_se_rm_a_lowerupper}
\title{Calculate SE RM for scenario A when f the manufacturer supplies lower and upper limits and coverage confidence interval (95 or 99...)}
\usage{
calculate_se_rm_a_lowerupper(upper, lower, coverage)
}
\arguments{
\item{upper}{upper limit}

\item{lower}{lower limit}

\item{coverage}{coverage}
}
\value{
SE RM
}
\description{
Calculate SE RM for scenario A when f the manufacturer supplies lower and upper limits and coverage confidence interval (95 or 99...)
}
