% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_se_rm.R
\name{calculate_se_rm_a_u}
\alias{calculate_se_rm_a_u}
\title{Calculate SE RM for scenario A when “standard error” or “standard uncertainty” (abbreviated by lowercase “u”) or “combined standard uncertainty” (often denoted by “uC ”)}
\usage{
calculate_se_rm_a_u(u)
}
\arguments{
\item{u}{“standard error” or “standard uncertainty” (abbreviated by lowercase “u”) or “combined standard uncertainty” (often denoted by “uC ”)}
}
\value{
SE RM
}
\description{
Calculate SE RM for scenario A when “standard error” or “standard uncertainty” (abbreviated by lowercase “u”) or “combined standard uncertainty” (often denoted by “uC ”)
}
