% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_se_rm.R
\name{calculate_se_rm_a_Ucoverage}
\alias{calculate_se_rm_a_Ucoverage}
\title{Calculate SE RM for scenario A when f the manufacturer supplies an “expanded uncertainty” (abbreviated by uppercase “U”) for the TV and coverage e.g. 95 or 99,}
\usage{
calculate_se_rm_a_Ucoverage(U, coverage)
}
\arguments{
\item{U}{expanded uncertainty}

\item{coverage}{coverage}
}
\value{
SE RM
}
\description{
Calculate SE RM for scenario A when f the manufacturer supplies an “expanded uncertainty” (abbreviated by uppercase “U”) for the TV and coverage e.g. 95 or 99,
}
