% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_se_rm.R
\name{calculate_se_rm_a_Uk}
\alias{calculate_se_rm_a_Uk}
\title{Calculate SE RM for scenario A when f the manufacturer supplies an “expanded uncertainty” (abbreviated by uppercase “U”) for the TV and the “coverage factor” (abbreviated by “k”)}
\usage{
calculate_se_rm_a_Uk(U, k)
}
\arguments{
\item{U}{expanded uncertainty}

\item{k}{coverage factor}
}
\value{
SE RM
}
\description{
Calculate SE RM for scenario A when f the manufacturer supplies an “expanded uncertainty” (abbreviated by uppercase “U”) for the TV and the “coverage factor” (abbreviated by “k”)
}
