% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_se_rm.R
\name{calculate_se_rm_scenario_b_c}
\alias{calculate_se_rm_scenario_b_c}
\title{Calculate SE RM for scenario B or C If the reference material has a TV determined by PT or peer group results}
\usage{
calculate_se_rm_scenario_b_c(sd_rm, nlab)
}
\arguments{
\item{sd_rm}{SD RM}

\item{nlab}{number of lab or peer group results}
}
\value{
SE RM
}
\description{
Calculate SE RM for scenario B or C If the reference material has a TV determined by PT or peer group results
}
