% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_se_rm.R
\name{calculate_se_rm_scenario_d_e}
\alias{calculate_se_rm_scenario_d_e}
\title{Calculate SE RM for scenario D or E If the TV represents a conventional quantity value or When working with a commercial QC material supplied with a TV for which the standard error cannot be estimated}
\usage{
calculate_se_rm_scenario_d_e()
}
\value{
SE RM
}
\description{
Calculate SE RM for scenario D or E If the TV represents a conventional quantity value or When working with a commercial QC material supplied with a TV for which the standard error cannot be estimated
}
