% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_se_x.R
\name{calculate_se_x}
\alias{calculate_se_x}
\title{Calculate SE x}
\usage{
calculate_se_x(nrun, nrep, SWL, SR)
}
\arguments{
\item{nrun}{Run number}

\item{nrep}{Number of repetitions per run n0}

\item{SWL}{SWL from aov table}

\item{SR}{SR from aov table}
}
\value{
SE X
}
\description{
Calculate SE x
}
