% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_ep_15.R
\name{create_table_ep_15}
\alias{create_table_ep_15}
\title{Create table for precision calculations}
\usage{
create_table_ep_15(data, data_type = "wider")
}
\arguments{
\item{data}{a long or a wider data.frame with the same structure of CLSIEP15::ferritin_long or CLSIEP15::ferritin_wider}

\item{data_type}{c('wider', 'long')}
}
\value{
a data.frame with renamed columns and structure adjustments
}
\description{
Create table for precision calculations
}
\examples{
data <- create_table_ep_15(ferritin_long, data_type = "longer")
}
