% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfc_references}
\alias{dfc_references}
\title{Reference of degrees of freedon based on tau given in the CLSI Manual}
\format{
`dfc_references`
A data frame with 390 rows and 4 columns:
\describe{
  \item{tau}{tau}
  \item{df}{degrees of freedon}
  \item{labs}{number of labs or peers}
  \item{runs}{number of runs}
  ...
}
}
\source{
CLSI EP15-A3
}
\usage{
dfc_references
}
\description{
Reference of degrees of freedon based on tau given in the CLSI Manual
}
\keyword{datasets}
