% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferritin_long}
\alias{ferritin_long}
\title{Ferrtin data used in CLSI document examples in wide format}
\format{
`ferritin_long`
A data frame with 25 rows and 3 columns:
\describe{
  \item{rep}{Repetition of sample}
  \item{name}{Run of the Runs obtained from 5 distinct days}
  \item{value}{result of the observation}
  ...
}
}
\source{
CLSI EP15-A3
}
\usage{
ferritin_long
}
\description{
Ferrtin data used in CLSI document examples in wide format
}
\keyword{datasets}
