% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferritin_wider}
\alias{ferritin_wider}
\title{Ferrtin data used in CLSI document examples in wide format}
\format{
`ferritin_wider`
A data frame with 5 rows and 6 columns:
\describe{
  \item{rep}{Repetition of sample}
  \item{Run_1, Run_2, Run_3, Run_4, Run_5}{Runs from 5 distinct days}
  ...
}
}
\source{
CLSI EP15-A3
}
\usage{
ferritin_wider
}
\description{
Ferrtin data used in CLSI document examples in wide format
}
\keyword{datasets}
